Move IT wiremock management into standalone class
In preparation for making changes how we manage our wiremock instance during integration tests this commit moves wiremock management into a standalone class.
